test(helpers): keep the golden naming tests in helpers_test.go
